requirements.txt
a2a-sdk
httpx
uvicorn
openai
pydantic
python-dotenv
